+++ electron/data/defaultSettings.ts
