+++ electron/data/schema.ts
