+++ src/actions/colorTheme.ts
@@ -1,6 +1,6 @@
 import { Action, ActionCreator, Dispatch } from "redux";
 import { Actions } from ".";
-import { IpcActions } from "../data/ipcActions";
+import { IpcActions } from "../shared/ipcActions";
 
 const { ipcRenderer } = window.require("electron");
