+++ src/shared/ipcActions.ts
@@ -1,3 +1,8 @@
+// Without ejecting from Create-React-App, there is
+// apparently no way to import from outside src/...
+// ---------------------------------------------------
+// These are the only thing that is shared between main
+// process (electron/) and renderer process code (src/).
 export const enum IpcActions {
   LoadPersistedData = "LoadPersistedData",
   CheckIfMacOs = "CheckIfMacOs",
